pyproject.toml
[tool.poetry]
name = "fishka-backend"
version = "0.1.0"
description = ""
authors = ["Grzegorz Jaworski <gostekk94@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.*"
Django = "^3.1.2"
gunicorn = "^20.0.4"
uvicorn = "^0.12.2"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.11.1"
django-filter = "^2.4.0"
djangorestframework-simplejwt = "^4.4.0"
django-allauth = "^0.43.0"
drf-yasg = "^1.17.1"
packaging = "^20.4"
django-dry-rest-permissions = "^1.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pylint = "^2.6.0"
pylint_django = "^2.3.0"
pylint_requests = "^0.1.1"
pytest = "^6.1.1"
pytest-cov = "^2.10.1"
pytest-pythonpath = "^0.7.3"
pytest-django = "^4.1.0"
factory-boy = "^3.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
